requirements/base.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.12
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.1.1
django-allauth==0.50.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0

